+++ src/sanitize/base64.ts
@@ -1,6 +1,6 @@
 import { getSanitizerFn } from "./_helpers.js";
 
 /**
- * Strips all non-base64 characters from `str`.
+ * Removes all non-base64 characters from `str`.
  */
 export const sanitizeBase64 = getSanitizerFn(/[^a-zA-Z0-9+/=]/g);