+++ src/sanitize/jwt.test.ts
@@ -3,13 +3,13 @@ import {
   NULL_CONTROL_CHAR,
   ZERO_WIDTH_AND_CONTROL_CHAR_REGEX,
 } from "./_helpers.js";
-import { sanitizeToken } from "./token.js";
+import { sanitizeJWT } from "./jwt.js";
 
-describe("sanitize: token", () => {
-  test("removes all non-token characters", () => {
+describe("sanitize: JWT", () => {
+  test("removes all non-JWT characters", () => {
     const input = `ab c${ZERO_WIDTH_SPACE}1 23${NULL_CONTROL_CHAR}~!@#$%^&*()_+-={}|:"<>?[]\\;',./`;
-    const result = sanitizeToken(input);
-    expect(result).toBe("abc123+./");
+    const result = sanitizeJWT(input);
+    expect(result).toBe("abc123_-.");
     expect(result).not.toMatch(ZERO_WIDTH_AND_CONTROL_CHAR_REGEX);
   });
 });