+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+# [1.3.0-next.1](https://github.com/Nerdware-LLC/ts-string-helpers/compare/v1.2.1...v1.3.0-next.1) (2024-04-25)
+
+### Bug Fixes
+
+* add quantifier to ensure base64 str is at least 2 chars long ([cb6992c](https://github.com/Nerdware-LLC/ts-string-helpers/commit/cb6992c757837a9b0cfdd4c5bde4108cce1fdad1))
+
+### Features
+
+* add base64URL sanitizer+validator fns ([ca5f181](https://github.com/Nerdware-LLC/ts-string-helpers/commit/ca5f1814fe15aadd3c01d5ee5b2f6bf7378071cf))
+
 ## [1.2.5](https://github.com/Nerdware-LLC/ts-string-helpers/compare/v1.2.4...v1.2.5) (2024-04-22)
 
 ## [1.2.4](https://github.com/Nerdware-LLC/ts-string-helpers/compare/v1.2.3...v1.2.4) (2024-04-04)